requirements.txt
py-cord>=2.6.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
